move phone number validation to its own file
