+++ notifications_utils/recipients.py
@@ -1,28 +1,24 @@
 import csv
 import re
 import sys
-from collections import namedtuple
 from contextlib import suppress
 from functools import lru_cache
 from io import StringIO
 from itertools import islice
 from typing import Optional, cast
 
-import phonenumbers
-from flask import current_app
 from ordered_set import OrderedSet
 
 from notifications_utils.formatters import (
-    ALL_WHITESPACE,
     strip_all_whitespace,
     strip_and_remove_obscure_whitespace,
 )
 from notifications_utils.insensitive_dict import InsensitiveDict
-from notifications_utils.international_billing_rates import (
-    COUNTRY_PREFIXES,
-    INTERNATIONAL_BILLING_RATES,
-)
 from notifications_utils.recipient_validation.errors import InvalidEmailError, InvalidPhoneError, InvalidRecipientError
+from notifications_utils.recipient_validation.phone_number import (
+    validate_and_format_phone_number,
+    validate_phone_number,
+)
 from notifications_utils.recipient_validation.postal_address import (
     address_line_7_key,
     address_lines_1_to_6_and_postcode_keys,
@@ -32,8 +28,6 @@ from notifications_utils.template import BaseLetterTemplate, Template
 
 from . import EMAIL_REGEX_PATTERN, hostname_part, tld_part
 from .qr_code import QrCodeTooLong
-
-uk_prefix = "44"
 
 first_column_headings = {
     "email": ["email address"],
@@ -471,128 +465,6 @@ class Cell:
         return self.error not in {None, self.missing_field_error}
 
 
-def normalise_phone_number(number):
-    for character in ALL_WHITESPACE + "()-+":
-        number = number.replace(character, "")
-
-    try:
-        list(map(int, number))
-    except ValueError as e:
-        raise InvalidPhoneError("Mobile numbers can only include: 0 1 2 3 4 5 6 7 8 9 ( ) + -") from e
-
-    return number.lstrip("0")
-
-
-def is_uk_phone_number(number):
-    if number.startswith("0") and not number.startswith("00"):
-        return True
-
-    number = normalise_phone_number(number)
-
-    if number.startswith(uk_prefix) or (number.startswith("7") and len(number) < 11):
-        return True
-
-    return False
-
-
-international_phone_info = namedtuple(
-    "PhoneNumber",
-    [
-        "international",
-        "crown_dependency",
-        "country_prefix",
-        "billable_units",
-    ],
-)
-
-
-def get_international_phone_info(number):
-    number = validate_phone_number(number, international=True)
-    prefix = get_international_prefix(number)
-    crown_dependency = _is_a_crown_dependency_number(number)
-
-    return international_phone_info(
-        international=(prefix != uk_prefix or crown_dependency),
-        crown_dependency=crown_dependency,
-        country_prefix=prefix,
-        billable_units=get_billable_units_for_prefix(prefix),
-    )
-
-
-CROWN_DEPENDENCY_RANGES = ["7781", "7839", "7911", "7509", "7797", "7937", "7700", "7829", "7624", "7524", "7924"]
-
-
-def _is_a_crown_dependency_number(number):
-    num_in_crown_dependency_range = number[2:6] in CROWN_DEPENDENCY_RANGES
-    num_in_tv_range = number[2:9] == "7700900"
-
-    return num_in_crown_dependency_range and not num_in_tv_range
-
-
-def get_international_prefix(number):
-    return next((prefix for prefix in COUNTRY_PREFIXES if number.startswith(prefix)), None)
-
-
-def get_billable_units_for_prefix(prefix):
-    return INTERNATIONAL_BILLING_RATES[prefix]["billable_units"]
-
-
-def use_numeric_sender(number):
-    prefix = get_international_prefix(normalise_phone_number(number))
-    return INTERNATIONAL_BILLING_RATES[prefix]["attributes"]["alpha"] == "NO"
-
-
-def validate_uk_phone_number(number):
-    number = normalise_phone_number(number).lstrip(uk_prefix).lstrip("0")
-
-    if not number.startswith("7"):
-        raise InvalidPhoneError(
-            "This does not look like a UK mobile number – double check the mobile number you entered"
-        )
-
-    if len(number) > 10:
-        raise InvalidPhoneError("Mobile number is too long")
-
-    if len(number) < 10:
-        raise InvalidPhoneError("Mobile number is too short")
-
-    return f"{uk_prefix}{number}"
-
-
-def validate_phone_number(number, international=False):
-    if (not international) or is_uk_phone_number(number):
-        return validate_uk_phone_number(number)
-
-    number = normalise_phone_number(number)
-
-    if len(number) < 8:
-        raise InvalidPhoneError("Mobile number is too short")
-
-    if len(number) > 15:
-        raise InvalidPhoneError("Mobile number is too long")
-
-    if get_international_prefix(number) is None:
-        raise InvalidPhoneError("Country code not found - double check the mobile number you entered")
-
-    return number
-
-
-validate_and_format_phone_number = validate_phone_number
-
-
-def try_validate_and_format_phone_number(number, international=None, log_msg=None):
-    """
-    For use in places where you shouldn't error if the phone number is invalid - for example if firetext pass us
-    something in
-    """
-    try:
-        return validate_and_format_phone_number(number, international)
-    except InvalidPhoneError as exc:
-        if log_msg:
-            current_app.logger.warning("%s: %s", log_msg, exc)
-        return number
-
-
 def validate_email_address(email_address):  # noqa (C901 too complex)
     # almost exactly the same as by https://github.com/wtforms/wtforms/blob/master/wtforms/validators.py,
     # with minor tweaks for SES compatibility - to avoid complications we are a lot stricter with the local part
@@ -653,24 +525,6 @@ def format_recipient(recipient):
     with suppress(InvalidEmailError):
         return validate_and_format_email_address(recipient)
     return recipient
-
-
-def format_phone_number_human_readable(phone_number):
-    try:
-        phone_number = validate_phone_number(phone_number, international=True)
-    except InvalidPhoneError:
-        # if there was a validation error, we want to shortcut out here, but still display the number on the front end
-        return phone_number
-    international_phone_info = get_international_phone_info(phone_number)
-
-    return phonenumbers.format_number(
-        phonenumbers.parse("+" + phone_number, None),
-        (
-            phonenumbers.PhoneNumberFormat.INTERNATIONAL
-            if international_phone_info.international
-            else phonenumbers.PhoneNumberFormat.NATIONAL
-        ),
-    )
 
 
 def allowed_to_send_to(recipient, allowlist):
