+++ tests/test_international_billing_rates.py
@@ -4,7 +4,7 @@ from notifications_utils.international_billing_rates import (
     COUNTRY_PREFIXES,
     INTERNATIONAL_BILLING_RATES,
 )
-from notifications_utils.recipients import use_numeric_sender
+from notifications_utils.recipient_validation.phone_number import use_numeric_sender
 
 
 def test_international_billing_rates_exists():
